+++ CHANGELOG.md
@@ -4,12 +4,13 @@ v0.4.9 / unreleased
   * Skip FormAuthenticityToken check for GET forms
   * Skip FormAuthenticityToken for action which is not relative path
   * Fix FormAction to not report offenses on valid scenarios
-  * UnusedAssign will not automatically remove assign (it changes logic when working with arrays/hash)
+  * UnusedAssign will not automatically remove assign if it might change the business logic (which is a scenario when filters modifying objects are used)
   * UnusedAssign will automatically rename result of background tag if variable not used
   * Fix reporting UndefinedObject's missing argument offenses when the same partial is used multiple times (previously offenses where displayed only for the last render)
   * Add autocorrector for UndefinedObject's missing argument error (explicitly provide null)
   * Add autocorrector for ImgLazyLoading
   * ConvertIncludeToRender will not report offense as autocorrect
+  * Improve inline GraphQL syntax check to raise error if result variable not provided
 
 v0.4.8 / 2023-12-20
 ==================